renamed/moved noteFields dir * now, fields stored in components > fields
