+++ src/components/fields/SymptomsCheckboxes.js
