+++ src/components/notes/NoteForm.js
@@ -1,22 +1,22 @@
-import { ChiefComplaintTextBox } from "./noteFields/ChiefComplaintTextBox";
-import { EducationLevelDropdown } from "./noteFields/EducationLevelDropdown";
-import { FamilySuicideButtons } from "./noteFields/FamilySuicideButtons";
-import { FinancialIssuesButtons } from "./noteFields/FinancialIssuesButtons";
-import { HeadInjuryDropdown } from "./noteFields/HeadInjuryDropdown";
-import { HousingStatusDropdown } from "./noteFields/HousingStatusDropdown";
-import { LegalIssuesButtons } from "./noteFields/LegalIssuesButtons";
-import { MaritalStatusDropdown } from "./noteFields/MaritalStatusDropdown";
-import { ModifyingFactorsTextBoxes } from "./noteFields/ModifyingFactorsTextBoxes";
-import { OccupationTextBox } from "./noteFields/OccupationTextBox";
-import { PatientAgeField } from "./noteFields/PatientAgeField";
-import { PatientGenderButtons } from "./noteFields/PatientGenderButtons";
-import { ReligiousButtons } from "./noteFields/ReligiousButtons";
-import { SymptomsCheckboxes } from "./noteFields/SymptomsCheckboxes";
-import { TraumaHistoryButtons } from "./noteFields/TraumaHistoryButtons";
-import { VeteranStatusDropdown } from "./noteFields/VeteranStatusDropdown";
-import { VisitDateAndTimeSelectors } from "./noteFields/VisitDateAndTimeSelectors";
-import { VisitLocationDropdown } from "./noteFields/VisitLocationDropdown";
-import { VisitTypeButtons } from "./noteFields/VisitTypeButtons";
+import { ChiefComplaintTextBox } from "../fields/ChiefComplaintTextBox";
+import { EducationLevelDropdown } from "../fields/EducationLevelDropdown";
+import { FamilySuicideButtons } from "../fields/FamilySuicideButtons";
+import { FinancialIssuesButtons } from "../fields/FinancialIssuesButtons";
+import { HeadInjuryDropdown } from "../fields/HeadInjuryDropdown";
+import { HousingStatusDropdown } from "../fields/HousingStatusDropdown";
+import { LegalIssuesButtons } from "../fields/LegalIssuesButtons";
+import { MaritalStatusDropdown } from "../fields/MaritalStatusDropdown";
+import { ModifyingFactorsTextBoxes } from "../fields/ModifyingFactorsTextBoxes";
+import { OccupationTextBox } from "../fields/OccupationTextBox";
+import { PatientAgeField } from "../fields/PatientAgeField";
+import { PatientGenderButtons } from "../fields/PatientGenderButtons";
+import { ReligiousButtons } from "../fields/ReligiousButtons";
+import { SymptomsCheckboxes } from "../fields/SymptomsCheckboxes";
+import { TraumaHistoryButtons } from "../fields/TraumaHistoryButtons";
+import { VeteranStatusDropdown } from "../fields/VeteranStatusDropdown";
+import { VisitDateAndTimeSelectors } from "../fields/VisitDateAndTimeSelectors";
+import { VisitLocationDropdown } from "../fields/VisitLocationDropdown";
+import { VisitTypeButtons } from "../fields/VisitTypeButtons";
 
 export const NoteForm = () => {
   return (
